labs/week1/practice/nguyen/chat-box-demo/requirements.txt
openai>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.40.0
langgraph>=0.6.6
langchain[openai]>=0.3.27
langgraph-checkpoint-sqlite>=2.0.0
aiosqlite>=0.21.0
pydantic>=2.0.0
grandalf
nest_asyncio>=1.6.0